Scripts/requirements.txt
meraki~=1.27.0
python-dotenv~=1.0.0
customtkinter~=5.1.2
requests~=2.28.2
elasticsearch~=8.6.2
matplotlib~=3.7.1
numpy~=1.24.3